pyproject.toml
[project]
name = "smartresearcher"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.12",
    "duckduckgo-search>=7.3.2",
    "langgraph>=0.2.70",
    "langsmith>=0.3.7",
    "openai>=1.61.1",
    "python-dotenv==1.0.1",
]
